pyproject.toml
[project]
name = "star-schema-olap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary,pool]>=3.3.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
